successfully addded feedback form

+++ ServerSide/UploadImage/imgur.js
@@ -12,13 +12,14 @@ async function uploadFile(localPath) {
 
         console.log(`Image ${localPath} uploaded successfully!`);
         //console.log(json);
-
-        fs.unlinkSync(localPath); // Deleting the file from the server
+        if(fs.existsSync(localPath))
+            fs.unlinkSync(localPath); // Deleting the file from the server
 
         return { id: json.data.id, link: json.data.link, deletehash: json.data.deletehash };
     } catch (err) {
         console.error('Error uploading image to Imgur:', err.message);
-        fs.unlinkSync(localPath); // Deleting the file from the server
+        if(fs.existsSync(localPath))
+            fs.unlinkSync(localPath); // Deleting the file from the server
         return { id: null, link: null, deletehash: null };
     }
 }
